move ScalarNumeric out of tdigest specific package
